Refactor into mock db class
