+++ test/HealthcareUnitTests.cpp
@@ -6,26 +6,8 @@
 #include <bsoncxx/builder/stream/document.hpp>
 #include <bsoncxx/json.hpp>
 
-#include "DatabaseManager.h"
 #include "Healthcare.h"
-
-class MockDatabaseManager : public DatabaseManager {
- public:
-  MockDatabaseManager() : DatabaseManager("mongodb://localhost:27017", true) {}
-
-  MOCK_METHOD(
-      void, findCollection,
-      (const std::string& collectionName,
-       (const std::vector<std::pair<std::string, std::string>>& keyValues),
-       (std::vector<bsoncxx::document::value> & result)),
-      (override));
-
-  MOCK_METHOD(
-      void, insertResource,
-      (const std::string& collectionName,
-       (const std::vector<std::pair<std::string, std::string>>& keyValues)),
-      (override));
-};
+#include "MockDatabaseManager.h"
 
 class HealthcareServiceUnitTests : public ::testing::Test {
  protected:
